refactor: NavigationBar
