+++ src/components/Home/Home.tsx
@@ -2,7 +2,6 @@ import React, { FunctionComponent } from "react";
 import { Redirect } from "react-router-dom";
 import { useConfigContext } from "../../common/context/config";
 import { usePersistedConfigFile } from "../../common/hook/usePersistedConfigFile";
-import { NavigationBar } from "../NavigationBar";
 import { ConfigFileDropZoneContainer } from "./ConfigFileDropZone";
 import { WalletDecryptionContainer } from "./WalletDecryption/WalletDecryptionContainer";
 
@@ -13,10 +12,5 @@ export const HomeContainer: FunctionComponent = () => {
   // If wallet has been decrypted, redirect to forms
   if (config) return <Redirect to="/forms-selection" />;
 
-  return (
-    <>
-      <NavigationBar />
-      {configFile ? <WalletDecryptionContainer /> : <ConfigFileDropZoneContainer />}
-    </>
-  );
+  return <>{configFile ? <WalletDecryptionContainer /> : <ConfigFileDropZoneContainer />}</>;
 };
