+++ src/components/PublishContainer/PublishPage/PublishPage.tsx
@@ -3,7 +3,6 @@ import { Redirect } from "react-router-dom";
 import { useFormsContext } from "../../../common/context/forms";
 import { usePublishQueue } from "../../../common/hook/usePublishQueue";
 import { Config } from "../../../types";
-import { NavigationBar } from "../../NavigationBar";
 import { PublishedScreen } from "../PublishedScreen";
 import { PublishErrorScreen } from "../PublishErrorScreen";
 
@@ -31,7 +30,6 @@ export const PublishPage: FunctionComponent<PublishPage> = ({ config }) => {
 
   return (
     <>
-      <NavigationBar />
       {publishState === "ERROR" ? (
         <PublishErrorScreen error={error} />
       ) : (
